Remove test that would enforce a way of scanning values into an *int
The reason for removing this test is that PGx4 and PGx5 don't agree
on how to work with *int, thus, if I made the behavior of KSQL
match pgx5 the test for pgx4 would break and vice-versa.

Fyi KSQL was made to do the same as pgx4 for now, I will create
an issue on pgx to ask why pgx5 is doing it differently, maybe
if he has a good reason for this I should do the same.