Fix small internal bug on how the StructInfo.NumFields() work
This bug was never noticed before because we actually are not using
this `NumFields()` function anywhere except on the new `kbuilder`
package, where we noticed this issue.

To be clear the issue was that if a struct had 4 fields, but only
the last 2 of them had `ksql` tags, the function `NumFields()` would
return the number 2, meaning that iterating over the fields using:

```go
for i := 0; i < info.NumFields(); i++ {
  field := info.ByIndex(i)
  fmt.Println(field.Valid)
  // do something
}
```

Would print `false` exactly two times. Because only index 2 and 3
actually had the tags.